+++ src/pages/setting/Tag_managment.jsx
@@ -51,14 +51,8 @@ const Tag_managment = () => {
 
   const setPermissionPage = async () => {
     const { user } = User.getCookieData();
-    // const { permissions } = Permission.convertPermissionValuesToBoolean([user]);
-    const mock_tag_permission = {
-      create: true,
-      delete: true,
-      update: true,
-      view: true,
-    };
-    setPagePermission(mock_tag_permission);
+    const { permissions } = Permission.convertPermissionValuesToBoolean([user]);
+    setPagePermission(permissions.user);
   };
 
   //Category function
@@ -81,31 +75,44 @@ const Tag_managment = () => {
   };
 
   const handleDeleteCategoryTag = async (items) => {
-    try {
-      const data = await User.deleteTagCategory(items.TagCategoryID, token);
-      if (data.code !== 404) {
-        Swal.fire({
-          icon: "success",
-          title: "Delete Tag Category Success ...",
-          text: `ลบ Tag Category ${items.TagCategoryName} สำเร็จ!`,
-        }).then((result) => {
-          if (
-            result.isConfirmed ||
-            result.dismiss === Swal.DismissReason.backdrop
-          ) {
-            getCategoryTag();
+    Swal.fire({
+      title: "คุณต้องการลบ Tag?",
+      text: `คุณต้องการ Tag : ${items.TagCategoryName}?`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      confirmButtonText: "ลบข้อมูล",
+      cancelButtonText: "ยกเลิก",
+      reverseButtons: true,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const data = await User.deleteTagCategory(items.TagCategoryID, token);
+          if (data.code !== 404) {
+            Swal.fire({
+              icon: "success",
+              title: "Delete Tag Category Success ...",
+              text: `ลบ Tag Category ${items.TagCategoryName} สำเร็จ!`,
+            }).then((result) => {
+              if (
+                result.isConfirmed ||
+                result.dismiss === Swal.DismissReason.backdrop
+              ) {
+                getCategoryTag();
+              }
+            });
+          } else {
+            Swal.fire({
+              icon: "error",
+              title: "เกิดข้อผิดพลาด!",
+              text: data.message,
+            });
           }
-        });
-      } else {
-        Swal.fire({
-          icon: "error",
-          title: "เกิดข้อผิดพลาด!",
-          text: data.message,
-        });
+        } catch (error) {
+          console.error("Error:", error);
+        }
       }
-    } catch (error) {
-      console.error("Error:", error);
-    }
+    });
   };
 
   // Tag Function
@@ -146,31 +153,44 @@ const Tag_managment = () => {
   };
 
   const removeTag = async (items) => {
-    try {
-      const data = await User.deleteTag(items.TagID, token);
-      if (data.code !== 404) {
-        Swal.fire({
-          icon: "success",
-          title: "Delete Tag Success ...",
-          text: `ลบ Tag สำเร็จ!`,
-        }).then((result) => {
-          if (
-            result.isConfirmed ||
-            result.dismiss === Swal.DismissReason.backdrop
-          ) {
-            getTagData(select_cat);
+    Swal.fire({
+      title: "คุณต้องการลบ Tag Option?",
+      text: `คุณต้องการลบ Tag Option : ${items.TagName}?`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      confirmButtonText: "ลบข้อมูล",
+      cancelButtonText: "ยกเลิก",
+      reverseButtons: true,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          const data = await User.deleteTag(items.TagID, token);
+          if (data.code !== 404) {
+            Swal.fire({
+              icon: "success",
+              title: "ลบ Tag Option Success ...",
+              text: `ลบ Tag Option ${items.TagName} สำเร็จ!`,
+            }).then((result) => {
+              if (
+                result.isConfirmed ||
+                result.dismiss === Swal.DismissReason.backdrop
+              ) {
+                getTagData(select_cat);
+              }
+            });
+          } else {
+            Swal.fire({
+              icon: "error",
+              title: "เกิดข้อผิดพลาด!",
+              text: data.message,
+            });
           }
-        });
-      } else {
-        Swal.fire({
-          icon: "error",
-          title: "เกิดข้อผิดพลาด!",
-          text: data.message,
-        });
+        } catch (error) {
+          console.log("error", error);
+        }
       }
-    } catch (error) {
-      console.log("error", error);
-    }
+    });
   };
 
   const handleEditTag = (items) => {
@@ -223,16 +243,18 @@ const Tag_managment = () => {
           {/* Left Panel */}
           <div className="bg-[#E8E8E8] col-span-2 h-[800px]">
             <div className="p-3">
-              <div className="font-poppins font-bold text-2xl">Category</div>
-              <div className="w-[150px] h-[40px] mt-3 bg-[#6425FE]  hover:bg-[#3b1694] text-white font-poppins flex justify-center items-center rounded-lg">
-                {page_permission.create ? (
-                  <button onClick={() => handleNewTagCategory()}>
-                    New Category +
-                  </button>
-                ) : (
-                  ""
-                )}
-              </div>
+              <div className="font-poppins font-bold text-2xl">Tag</div>
+              {page_permission.create ? (
+                <div
+                  onClick={() => handleNewTagCategory()}
+                  className="w-[150px] h-[40px] mt-3 bg-[#6425FE]  hover:bg-[#3b1694] text-white font-poppins flex justify-center items-center rounded-lg cursor-pointer"
+                >
+                  <button>New Tag +</button>
+                </div>
+              ) : (
+                ""
+              )}
+
               <div className="h-[700px] overflow-y-auto">
                 {category_data.length > 0 &&
                   category_data.map((items, key) => (
@@ -242,7 +264,7 @@ const Tag_managment = () => {
                         className={`grid grid-cols-7 gap-2 mt-5 
                   ${
                     items.TagCategoryID === select_cat.TagCategoryID
-                      ? "text-[#6425FE]"
+                      ? "text-white bg-[#6425FE] h-[55px] border rounded-lg"
                       : ""
                   } 
                   cursor-pointer`}
@@ -252,30 +274,50 @@ const Tag_managment = () => {
                           <div className={`font-poppins text-2xl`}>
                             {items.TagCategoryName}
                           </div>
-                          <div className="text-xs">{items.TagDesc}</div>
+                          <div className="text-xs">
+                            {items.TagDesc || "No Description"}
+                          </div>
                         </div>
 
                         <div className="col-span-2">
                           <div className="flex justify-center items-center mt-3 space-x-4">
-                            <button
-                              onClick={() =>
-                                setModalCreateNewCategory(
-                                  !modalCreateNewCategory
-                                )
-                              }
-                            >
-                              <RiEditLine
-                                size={20}
-                                className="text-[#6425FE] hover:text-[#3b1694]"
-                              />
-                            </button>
-                            <button>
-                              <RiDeleteBin5Line
-                                onClick={() => handleDeleteCategoryTag(items)}
-                                size={20}
-                                className="text-[#6425FE] hover:text-[#3b1694]"
-                              />
-                            </button>
+                            {page_permission.update ? (
+                              <button
+                                onClick={() =>
+                                  setModalCreateNewCategory(
+                                    !modalCreateNewCategory
+                                  )
+                                }
+                              >
+                                <RiEditLine
+                                  size={20}
+                                  className={`${
+                                    items.TagCategoryID ===
+                                    select_cat.TagCategoryID
+                                      ? "text-white hover:text-gray-500"
+                                      : "text-[#6425FE] hover:text-[#3b1694]"
+                                  }`}
+                                />
+                              </button>
+                            ) : (
+                              <></>
+                            )}
+                            {page_permission.delete ? (
+                              <button>
+                                <RiDeleteBin5Line
+                                  onClick={() => handleDeleteCategoryTag(items)}
+                                  size={20}
+                                  className={`${
+                                    items.TagCategoryID ===
+                                    select_cat.TagCategoryID
+                                      ? "text-white hover:text-gray-500"
+                                      : "text-[#6425FE] hover:text-[#3b1694]"
+                                  }`}
+                                />
+                              </button>
+                            ) : (
+                              <></>
+                            )}
                           </div>
                         </div>
                       </div>
@@ -289,7 +331,7 @@ const Tag_managment = () => {
           <div className="col-span-5 bg-[#FAFAFA] w-full">
             <div className="p-3">
               <div className="font-poppins font-bold text-2xl">
-                Tag : {select_cat.TagCategoryName}
+                Tag Options : {select_cat.TagCategoryName}
               </div>
 
               <div className="grid grid-cols-7 gap-2 mt-5">
@@ -297,19 +339,28 @@ const Tag_managment = () => {
                   <input
                     type="text"
                     className="w-[100%] h-[100%] border border-gray-300 rounded-md pl-4 placeholder-ml-2"
-                    placeholder="Enter Tag"
+                    placeholder="กรอกชื่อ Tag Option"
                     value={new_tag}
                     onChange={(e) => setNewTag(e.target.value)}
+                    disabled={
+                      page_permission.create || page_permission.update
+                        ? false
+                        : true
+                    }
                   />
                 </div>
-                <div className="col-span-1 h-12">
-                  <button
-                    onClick={() => addNewTag()}
-                    className="w-[100%] h-[100%] rounded-lg bg-[#6425FE]  hover:bg-[#3b1694] font-poppins text-white"
-                  >
-                    Add +
-                  </button>
-                </div>
+                {page_permission.create || page_permission.update ? (
+                  <div className="col-span-1 h-12">
+                    <button
+                      onClick={() => addNewTag()}
+                      className="w-[100%] h-[100%] rounded-lg bg-[#6425FE]  hover:bg-[#3b1694] font-poppins text-white"
+                    >
+                      Add Tag Option
+                    </button>
+                  </div>
+                ) : (
+                  <></>
+                )}
               </div>
               <div className="col-span-5 mt-5 h-[600px] overflow-y-auto">
                 <div className="flex flex-wrap">
@@ -322,18 +373,29 @@ const Tag_managment = () => {
                           flexBasis: `calc(30% - 5px)`, // Increased width and adjusted spacing
                         }}
                       >
-                        <div className="flex justify-center items-center mr-1 ml-1">
-                          <IoIosClose
-                            onClick={() => removeTag(items)}
-                            className="text-[#6425FE] hover:text-[#3b1694] cursor-pointer"
-                          />
-                        </div>
-                        <div
-                          onClick={() => handleEditTag(items)}
-                          className="flex-grow lg:text-sm md:text-xs font-poppins flex justify-center cursor-pointer"
-                        >
-                          {items.TagName}
-                        </div>
+                        {page_permission.delete ? (
+                          <div className="flex justify-center items-center mr-1 ml-1">
+                            <IoIosClose
+                              onClick={() => removeTag(items)}
+                              className="text-[#6425FE] hover:text-[#3b1694] cursor-pointer"
+                            />
+                          </div>
+                        ) : (
+                          <></>
+                        )}
+
+                        {page_permission.update ? (
+                          <div
+                            onClick={() => handleEditTag(items)}
+                            className="flex-grow lg:text-sm md:text-xs font-poppins flex justify-center cursor-pointer"
+                          >
+                            {items.TagName}
+                          </div>
+                        ) : (
+                          <div className="flex-grow lg:text-sm md:text-xs font-poppins flex justify-center ">
+                            {items.TagName}
+                          </div>
+                        )}
                       </div>
                     ))
                   ) : (
@@ -343,36 +405,6 @@ const Tag_managment = () => {
                   )}
                 </div>
               </div>
-              {/* {tag_data.length > 0 ? (
-                <div className="grid grid-cols-5 gap-2 mt-5">
-                  {tag_data.map((items, index) => (
-                    <div key={items.TagID}>
-                      <div className="relative w-[150px] lg:w-[180px] h-[40px] flex items-center justify-center  ml-3 border border-gray-300 hover:border-[#6425FE] rounded-full">
-                        <div className=" absolute inset-y-0 left-0 flex items-center pl-2 ">
-                          <button onClick={() => removeTag(items)}>
-                            <IoIosClose
-                              size="22"
-                              className="text-[#6425FE] hover:text-[#504f53] cursor-pointer"
-                            />
-                          </button>
-                        </div>
-                        <div
-                          onClick={() => handleEditTag(items)}
-                          className="text-sm md:text-xs text-bold font-poppins cursor-pointer "
-                        >
-                          {items.TagName}
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              ) : (
-                <div className="p-4">
-                  <div className="flex justify-center items-center h-[650px]">
-                    <span className="text-gray-300 text-4xl">No Tag(s)</span>
-                  </div>
-                </div>
-              )} */}
             </div>
           </div>
           {/* Right Panel */}
@@ -391,6 +423,7 @@ const Tag_managment = () => {
           setModalCreateNewCategory={setModalCreateNewCategory}
           modalCreateNewCategory={modalCreateNewCategory}
           select_cat={select_cat}
+          getCategoryTag={getCategoryTag}
         />
       )}
